Update NixOS and home-manager to latest versions
This commit updates the NixOS version from 24.05 to 24.11, the home-manager version from release-24.05 to release-24.11, and modifies the
`threed-printing.nix` file to skip installing a broken package. Additionally, it updates the `hardware-configuration.nix` file to enable hardware
acceleration for NVIDIA graphics and disable sound. Finally, it adjusts the `system.nix` file to rename the `noto-fonts-cjk` package to
`noto-fonts-cjk-sans` and disables the Pipewire sound support.