Update flake.lock and modify kde.nix configuration
This commit updates the `flake.lock` file to reflect the latest changes in the Nix packages and repositories. The `flake.lock` file now contains updated
revision numbers for `nixos-hardware`, `nixpkgs`, `nixpkgs-unstable`, and `kde`. Additionally, the `kde.nix` configuration has been modified to enable more
services and adjust the timeout settings for `plasma-powerdevil`. This ensures that the system uses the latest dependencies and configurations while
maintaining compatibility and functionality.