Add backup file extension and update kernel packages in flake.nix
This commit adds the `backupFileExtension` option to the home-manager configuration in `flake.nix`, setting it to 'backup'. This ensures that any files backed up by home-manager have a '.backup' extension, improving clarity and organization. Additionally, the commit updates the kernel packages in the hardware-configuration.nix file to use the latest 6.12 version of the linuxPackages_6.12 package set, ensuring compatibility with the latest kernel features and security patches.