Update nixos-hardware repository hash and revision
This commit updates the `narHash` and `rev` fields for the `nixos-hardware` repository in the `flake.lock` file. The `lastModified` timestamp has been updated to reflect the latest changes, and the `sha256-hash` has been recalculated. This ensures that the lockfile remains accurate and up-to-date with the current state of the `nixos-hardware` repository.