Add esphome to developer tools
This change adds the ESPHome tool to the list of developer tools. ESPHome is a popular open-source project for building IoT devices using the ESP8266 and ESP32 chipsets. By including ESPHome in the list, it provides developers with a set of tools necessary for developing firmware for these devices, enhancing their development capabilities and expanding their range of supported projects.